refactor: extract sandbox path literals
